Derive setsquare classes from geometryTool classes
