+++ src/core/model/GeometryTool.cpp
@@ -1,0 +1,43 @@
+#include "GeometryTool.h"
+
+#include <initializer_list>  // for initializer_list
+
+GeometryTool::GeometryTool(double h, double r, double tx, double ty):
+        height(h),
+        rotation(r),
+        translationX(tx),
+        translationY(ty),
+        viewPool(std::make_shared<xoj::util::DispatchPool<xoj::view::GeometryToolView>>()),
+        handlerPool(std::make_shared<xoj::util::DispatchPool<GeometryToolHandler>>()) {}
+
+GeometryTool::~GeometryTool() {}
+
+void GeometryTool::setHeight(double height) { this->height = height; }
+auto GeometryTool::getHeight() const -> double { return this->height; }
+
+void GeometryTool::setRotation(double rotation) { this->rotation = rotation; }
+auto GeometryTool::getRotation() const -> double { return this->rotation; }
+
+void GeometryTool::setTranslationX(double x) { this->translationX = x; }
+auto GeometryTool::getTranslationX() const -> double { return this->translationX; }
+
+void GeometryTool::setTranslationY(double y) { this->translationY = y; }
+auto GeometryTool::getTranslationY() const -> double { return this->translationY; }
+
+void GeometryTool::getMatrix(cairo_matrix_t& matrix) const {
+    cairo_matrix_init_identity(&matrix);
+    cairo_matrix_translate(&matrix, this->translationX, this->translationY);
+    cairo_matrix_rotate(&matrix, this->rotation);
+    cairo_matrix_scale(&matrix, CM, CM);
+}
+
+auto GeometryTool::getViewPool() const -> const std::shared_ptr<xoj::util::DispatchPool<xoj::view::GeometryToolView>>& {
+    return viewPool;
+}
+auto GeometryTool::getHandlerPool() const -> const std::shared_ptr<xoj::util::DispatchPool<GeometryToolHandler>>& {
+    return handlerPool;
+}
+
+auto GeometryTool::getStroke() const -> Stroke* { return this->stroke; }
+
+void GeometryTool::setStroke(Stroke* s) { this->stroke = s; }
