+++ src/core/model/Setsquare.cpp
@@ -8,13 +8,7 @@
 
 Setsquare::Setsquare(): Setsquare(INITIAL_HEIGHT, .0, INITIAL_X, INITIAL_Y) {}
 
-Setsquare::Setsquare(double height, double rotation, double x, double y):
-        height(height),
-        rotation(rotation),
-        translationX(x),
-        translationY(y),
-        viewPool(std::make_shared<xoj::util::DispatchPool<xoj::view::SetsquareView>>()),
-        handlerPool(std::make_shared<xoj::util::DispatchPool<SetsquareInputHandler>>()) {}
+Setsquare::Setsquare(double height, double rotation, double x, double y): GeometryTool(height, rotation, x, y) {}
 
 Setsquare::~Setsquare() { viewPool->dispatchAndClear(xoj::view::SetsquareView::FINALIZATION_REQUEST, Range()); }
 
@@ -42,33 +36,3 @@ void Setsquare::notify() const {
     handlerPool->dispatch(SetsquareInputHandler::UPDATE_VALUES, this->getHeight(), this->getRotation(),
                           this->getTranslationX(), this->getTranslationY());
 }
-
-void Setsquare::setHeight(double height) { this->height = height; }
-auto Setsquare::getHeight() const -> double { return this->height; }
-
-void Setsquare::setRotation(double rotation) { this->rotation = rotation; }
-auto Setsquare::getRotation() const -> double { return this->rotation; }
-
-void Setsquare::setTranslationX(double x) { this->translationX = x; }
-auto Setsquare::getTranslationX() const -> double { return this->translationX; }
-
-void Setsquare::setTranslationY(double y) { this->translationY = y; }
-auto Setsquare::getTranslationY() const -> double { return this->translationY; }
-
-void Setsquare::getMatrix(cairo_matrix_t& matrix) const {
-    cairo_matrix_init_identity(&matrix);
-    cairo_matrix_translate(&matrix, this->translationX, this->translationY);
-    cairo_matrix_rotate(&matrix, this->rotation);
-    cairo_matrix_scale(&matrix, CM, CM);
-}
-
-auto Setsquare::getViewPool() const -> const std::shared_ptr<xoj::util::DispatchPool<xoj::view::SetsquareView>>& {
-    return viewPool;
-}
-auto Setsquare::getHandlerPool() const -> const std::shared_ptr<xoj::util::DispatchPool<SetsquareInputHandler>>& {
-    return handlerPool;
-}
-
-auto Setsquare::getStroke() const -> Stroke* { return this->stroke; }
-
-void Setsquare::setStroke(Stroke* s) { this->stroke = s; }
